backend/api_gateway_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
python-jose[cryptography]
